add more unit tests for process of selecting dedicated ALB
