+++ tests/integration/alb/test_alb_provisioning.py
@@ -39,51 +39,6 @@ from tests.integration.alb.test_alb_update import (
 # these subtasks when testing failure scenarios.
 
 
-def test_gets_lowest_used_alb(alb):
-    alb.expect_get_certificates_for_listener("listener-arn-0", 1)
-    alb.expect_get_listeners("listener-arn-0")
-    assert get_lowest_used_alb(["listener-arn-0"]) == (
-        "alb-listener-arn-0",
-        "listener-arn-0",
-    )
-    alb.expect_get_certificates_for_listener("listener-arn-0", 2)
-    alb.expect_get_certificates_for_listener("listener-arn-1", 1)
-    alb.expect_get_listeners("listener-arn-1")
-    assert get_lowest_used_alb(["listener-arn-0", "listener-arn-1"]) == (
-        "alb-listener-arn-1",
-        "listener-arn-1",
-    )
-    alb.expect_get_certificates_for_listener("listener-arn-1", 1)
-    alb.expect_get_certificates_for_listener("listener-arn-0", 2)
-    alb.expect_get_listeners("listener-arn-1")
-    assert get_lowest_used_alb(["listener-arn-1", "listener-arn-0"]) == (
-        "alb-listener-arn-1",
-        "listener-arn-1",
-    )
-    alb.expect_get_certificates_for_listener("listener-arn-1", 1)
-    alb.expect_get_certificates_for_listener("listener-arn-2", 2)
-    alb.expect_get_certificates_for_listener("listener-arn-0", 2)
-    alb.expect_get_listeners("listener-arn-1")
-    assert get_lowest_used_alb(
-        ["listener-arn-1", "listener-arn-2", "listener-arn-0"]
-    ) == ("alb-listener-arn-1", "listener-arn-1")
-    alb.expect_get_certificates_for_listener("listener-arn-0", 19)
-    alb.expect_get_certificates_for_listener("listener-arn-1", 0)
-    alb.expect_get_certificates_for_listener("listener-arn-2", 25)
-    alb.expect_get_certificates_for_listener("listener-arn-3", 20)
-    alb.expect_get_certificates_for_listener("listener-arn-4", 17)
-    alb.expect_get_listeners("listener-arn-1")
-    assert get_lowest_used_alb(
-        [
-            "listener-arn-0",
-            "listener-arn-1",
-            "listener-arn-2",
-            "listener-arn-3",
-            "listener-arn-4",
-        ]
-    ) == ("alb-listener-arn-1", "listener-arn-1")
-
-
 def test_provision_happy_path(
     client, dns, tasks, route53, iam_govcloud, simple_regex, alb
 ):
